Time: 103 ms (73.50%), Space: 61.3 MB (20.03%) - LeetHub

+++ 713-subarray-product-less-than-k/713-subarray-product-less-than-k.cpp
@@ -3,14 +3,16 @@ public:
     int numSubarrayProductLessThanK(vector<int>& nums, int k) {
         int ans=0;
         int i=0,j=0;
-        int pro=1;
+        long long int pro=1;
+        
         while(j<nums.size())
         {
             pro*=nums[j];
+            
             while(pro>=k && i<=j)
                 pro/=nums[i++];
+            
             ans+=j-i+1;
-            //cout<<ans<<" "<<i<<" "<<j<<endl;
             j++;
         }
         return ans;
